refactor: Unify emoji rendering into single Emoji component
Consolidate CustomEmoji and UnicodeEmoji into a single unified Emoji
component with a cleaner API:

- source: "unicode" | "custom" - determines rendering mode
- value: string - emoji char (unicode) or URL (custom)
- shortcode: string - for tooltips
- size: "xs" | "sm" | "md" | "lg" - consistent sizing
- showTooltip: boolean - optional tooltip on hover

Benefits:
- Single import instead of two components
- Maps directly to EmojiSearchResult shape
- Consistent API regardless of emoji type
- One place for sizing, tooltips, and error handling

Updated components:
- EmojiSuggestionList, EmojiPickerDialog
- MessageReactions, ReactionRenderer, ReactionCompactPreview
- EmojiListRenderer, EmojiSetRenderer, EmojiSetDetailRenderer
- RichText/Emoji

Removed:
- CustomEmoji.tsx (merged into Emoji.tsx)
- UnicodeEmoji.tsx (merged into Emoji.tsx)
