+++ src/components/nostr/kinds/ReactionRenderer.tsx
@@ -7,8 +7,7 @@ import { KindRenderer } from "./index";
 import { EventCardSkeleton } from "@/components/ui/skeleton";
 import { parseReplaceableAddress } from "applesauce-core/helpers/pointers";
 import { EMOJI_SHORTCODE_REGEX } from "@/lib/emoji-helpers";
-import { CustomEmoji } from "../CustomEmoji";
-import { UnicodeEmoji } from "../UnicodeEmoji";
+import { Emoji } from "../Emoji";
 
 /**
  * Renderer for Kind 7 - Reactions
@@ -102,7 +101,15 @@ export function Kind7Renderer({ event }: BaseEventProps) {
       case "😊":
         return <Smile className="size-4 fill-yellow-500 text-yellow-500" />;
       default:
-        return <UnicodeEmoji emoji={content} size="md" showTooltip={false} />;
+        return (
+          <Emoji
+            source="unicode"
+            value={content}
+            shortcode={content}
+            size="md"
+            showTooltip={false}
+          />
+        );
     }
   };
 
@@ -112,9 +119,10 @@ export function Kind7Renderer({ event }: BaseEventProps) {
         {/* Reaction indicator */}
         <div className="flex items-center gap-2">
           {parsedReaction.type === "custom" ? (
-            <CustomEmoji
+            <Emoji
+              source="custom"
+              value={parsedReaction.url}
               shortcode={parsedReaction.shortcode}
-              url={parsedReaction.url}
               size="md"
             />
           ) : (
